requirements.txt
requests==2.31.0
pandas==2.2.0
pandas-ta==0.4.71b0
python-binance==1.0.19
colorama==0.4.6
python-dotenv==1.0.0
